Disabled Authentication, able to add products in cart and check the list of products in cart

+++ src/main/java/com/extron/MyFirstECom/Repository/CartItemRepo.java
@@ -3,5 +3,5 @@ package com.extron.MyFirstECom.Repository;
 import com.extron.MyFirstECom.Model.CartItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartItemRepo extends JpaRepository<CartItem, Integer> {
+public interface CartItemRepo extends JpaRepository<CartItem, Long> {
 }
